Reorganized into a directory structure that is more suitable for large applications.
